Reorganize effects into individual files
