+++ src/index.ts
@@ -1,4 +1,5 @@
 export { Channel } from "./Channel";
-export { take, takeAdvanced } from "./effects";
+export { take } from "./effects/take";
+export { takeAdvanced } from "./effects/takeAdvanced";
 export { Store } from "./Store";
 export { IAction, IActionType, IChannel, IStore } from "./types";
